+++ src/main/java/com/codecool/web/service/CartService.java
@@ -16,7 +16,7 @@ public interface CartService {
 
     TotalDto getTotalDto(int id) throws SQLException, ServiceException;
 
-    int getTotalCartCost(int userId) throws SQLException;
+    double getTotalCartCost(int userId) throws SQLException;
 
     void deleteProductFromCart(int userId, int prodId) throws SQLException;
 }
