+++ src/main/java/com/codecool/web/service/simple/SimpleCartService.java
@@ -60,7 +60,7 @@ public class SimpleCartService implements CartService {
          }
     }
 
-    public int getTotalCartCost(int userId) throws SQLException {
+    public double getTotalCartCost(int userId) throws SQLException {
         return cartDao.getTotalCartCost(userId);
     }
